+++ front-angular/src/app/shared/models/types.ts
@@ -26,7 +26,7 @@ export interface Modulo {
   createdAt: string;
 }
 
-export type TipoPrograma = 'Menú' | 'Submenú' | 'Maestro' | 'Transacción' | 'Proceso' | 'Consulta' | 'Reporte' | 'Objeto';
+export type TipoPrograma = 'Menú' | 'Submenú' | 'Tapview' | 'Maestro' | 'Transacción' | 'Proceso' | 'Consulta' | 'Reporte' | 'Objeto';
 
 export interface Programa {
   id: string;
@@ -40,7 +40,7 @@ export interface Programa {
   createdAt: string;
 }
 
-export type TipoControl = 'Caja de Texto' | 'Botón' | 'Check' | 'Combo' | 'Grid' | 'Option' | 'Otros';
+export type TipoControl = 'Caja de Texto' | 'Botón' | 'Check' | 'Combo' | 'Grid' | 'Filtro' | 'Option' | 'Otros';
 
 export interface Control {
   id: string;